+++ web/app.js
@@ -85,7 +85,7 @@ import { Preferences } from "web-preferences";
 import { SecondaryToolbar } from "web-secondary_toolbar";
 import { Toolbar } from "web-toolbar";
 import { ViewHistory } from "./view_history.js";
-import {setupCustomCapabilities} from "./custom.js";
+import {setupCustomCapabilities} from "./custom/index.js";
 
 const FORCE_PAGES_LOADED_TIMEOUT = 10000; // ms
